requirements.txt
streamlit==1.32.0
pandas
pyodbc
sqlalchemy
azure-identity
plotly